feat: add ShortestPath() (Dijkstra), GoBFS(), GoDFS() graph methods

+++ graph/edge.go
@@ -9,17 +9,17 @@ type edge struct {
 	to string
 
 	// The edge weight value
-	weight int
+	weight float64
 }
 
 // newEdge creates a new edge initialized with given origin id, destination id,
 // and weight value.
-func newEdge(from, to string, weight int) *edge {
+func newEdge(from, to string, weight float64) *edge {
 	return &edge{from, to, weight}
 }
 
 // setWeight sets the weight of the current edge and returns the edge.
-func (e *edge) setWeight(v int) *edge {
+func (e *edge) setWeight(v float64) *edge {
 	e.weight = v
 	return e
 }
